Refactor TrackerCommandTest. Implement Delete and List functions for Tracker Command. Update manual.txt and UG with new Tracker Command syntax. Fix merge conflict.
