+++ src/main/java/oof/logic/command/productivity/exceptions/ScheduleEmptyException.java
@@ -1,9 +1,8 @@
-package oof.exception.command;
+package oof.logic.command.productivity.exceptions;
 
 import oof.commons.exceptions.command.CommandException;
 
 public class ScheduleEmptyException extends CommandException {
-
     public ScheduleEmptyException(String message) {
         super(message);
     }
