+++ src/test/java/oof/logic/command/TrackerCommandTest.java
@@ -1,9 +1,9 @@
-package oof.command;
+package oof.logic.command;
 
 import oof.Oof;
-import oof.exception.ParserException;
-import oof.exception.StorageFileCorruptedException;
-import oof.exception.command.CommandException;
+import oof.commons.exceptions.ParserException;
+import oof.commons.exceptions.StorageFileCorruptedException;
+import oof.commons.exceptions.command.CommandException;
 import oof.model.task.Task;
 import oof.model.task.TaskList;
 import oof.model.tracker.Tracker;
